+++ src/vscode/treeviews/Workspaces/FabricSqlEndpoint.ts
@@ -40,9 +40,23 @@ export class FabricSqlEndpoint extends FabricItem {
 		return children;
 	}
 
-	async syncMetadata(): Promise<void> {
-		// https://learn.microsoft.com/en-us/fabric/data-factory/pipeline-rest-api#run-on-demand-item-job
+	static async syncMetadata(sqlEndpoint: FabricSqlEndpoint): Promise<void> {
+		// https://learn.microsoft.com/en-us/rest/api/fabric/sqlendpoint/items/refresh-sql-endpoint-metadata?tabs=HTTP
 
+		const preview: boolean = true;
+		const endpoint = Helper.joinPath(sqlEndpoint.apiPath, `refreshMetadata?preview=${preview}`);
+
+		const response = await FabricApiService.post(endpoint, {}, { "raw": false, "awaitLongRunningOperation": false });
+
+		if (response.error) {
+			ThisExtension.Logger.logError(`Metadata Sync failed for SQL Endpoint '${sqlEndpoint.itemName}': ${response.error.message}`, true);
+		}
+		else {
+			ThisExtension.Logger.logInfo(`Metadata Sync started for SQL Endpoint '${sqlEndpoint.itemName}' ...`, 5000);
+		}
+	}
+	
+	async syncMetadataOld(): Promise<void> {
 		const endpoint = `https://api.powerbi.com/v1.0/myorg/lhdatamarts/${this.itemId}`;
 
 		const body = {